+++ src/openlayers/mapping/ImageSuperMapRest.js
@@ -18,16 +18,17 @@ import {Util} from '../core/Util';
  *        tileGrid - [{ol.tilegrid.TileGrid}]{@linkdoc-openlayers/ol.tilegrid.TileGrid} 瓦片网格对象。<br>
  *        serverType - {SuperMap.ServerType} 服务类型。<br>
  *        redirect - {boolean} 是否重定向，默认为false。<br>
- *        transparent - {boolean} 图片是否透明，默认为true。<br>
+ *        transparent - {boolean} 瓦片是否透明，默认为true。<br>
  *        cacheEnabled - {boolean} 是否使用服务端的缓存，默认为 true，即使用服务端的缓存。<br>
- *        prjCoordSys - {Object} 请求的地图的坐标参考系统。当此参数设置的坐标系统不同于地图的原有坐标系统时， 系统会进行动态投影，并返回动态投影后的地图图片。例如：{"epsgCode":3857}。<br>
+ *        prjCoordSys - {Object} 请求的地图的坐标参考系统。当此参数设置的坐标系统不同于地图的原有坐标系统时， 系统会进行动态投影，并返回动态投影后的地图瓦片。例如：{"epsgCode":3857}。<br>
  *        layersID - {string} 获取进行切片的地图图层 ID，即指定进行地图切片的图层，可以是临时图层集，也可以是当前地图中图层的组合。如果此参数缺省则对全部图层进行切片。layersID 可以是临时图层创建时templayers的ID。<br>
  *        clipRegionEnabled - {boolean} 地图显示裁剪的区域。是一个面对象，当 clipRegionEnabled = true 时有效，即地图只显示该区域覆盖的部分。<br>
  *        clipRegion - [{ol.geom.Geometry}]{@linkdoc-openlayers/ol.geom.Geometry} 地图显示裁剪的区域。是一个面对象，当 clipRegionEnabled = true 时有效，即地图只显示该区域覆盖的部分。<br>
  *        overlapDisplayed - {boolean} 地图对象在同一范围内时，是否重叠显示，默认为 false。如果为 true，则同一范围内的对象会直接压盖；如果为 false 则通过 overlapDisplayedOptions 控制对象不压盖显示。<br>
  *        overlapDisplayedOptions - {@link SuperMap.OverlapDisplayedOptions} 避免地图对象压盖显示的过滤选项，当 overlapDisplayed 为 false 时有效，用来增强对地图对象压盖时的处理。<br>
  *        tileversion - {string} 切片版本名称，_cache 为 true 时有效。<br>
- *        tileProxy - {string} 代理地址
+ *        tileProxy - {string} 代理地址。
+ *        format - {string} 瓦片表述类型，支持 "png" 、"bmp" 、"jpg" 和 "git" 四种表述类型，默认为 "png"。
  * @extends ol.source.TileImage{@linkdoc-openlayers/ol.source.TileImage}
  */
 export class ImageSuperMapRest extends ol.source.TileImage {
@@ -40,7 +41,10 @@ export class ImageSuperMapRest extends ol.source.TileImage {
             new ol.Attribution({
                 html: "Map Data <span>© <a href='http://support.supermap.com.cn/product/iServer.aspx' target='_blank'>SuperMap iServer</a></span> with <a href='http://icltest.supermapol.com/'>© SuperMap iClient</a>"
             });
-        var layerUrl = options.url + "/image.png?";
+
+        options.format = options.format ? options.format : "png";
+        var layerUrl = options.url + "/image." + options.format + "?";
+
         options.serverType = options.serverType || ServerType.ISERVER;
         //为url添加安全认证信息片段
         layerUrl = appendCredential(options.url, layerUrl, options.serverType);